feat(ui-design): Get the directory structure up and running
