+++ config/webpack.common.js
@@ -7,7 +7,7 @@ var extractTextPlugin = require('extract-text-webpack-plugin');
 var webpackNotifierPlugin = require('webpack-notifier');
 
 const METADATA = {
-	title: 'StepChange Angular 2 Application Seed'
+	title: 'StepChange Deadpool'
 }
 
 module.exports = {
