+++ package.json
@@ -50,10 +50,10 @@
     "@angular/http": "2.0.0-rc.4",
     "@angular/platform-browser": "2.0.0-rc.4",
     "@angular/platform-browser-dynamic": "2.0.0-rc.4",
-    "@angular2-material/icon": "2.0.0-alpha.6-2",
+    "@angular/router": "^3.0.0-alpha.8",
     "core-js": "^2.4.0",
     "raven-js": "^3.3.0",
-    "reflect-metadata": "^0.1.3",
+    "reflect-metadata": "0.1.3",
     "rxjs": "5.0.0-beta.6",
     "zone.js": "0.6.12"
   },
