--- src/app/app.scss
@@ -1,6 +1,0 @@
-.title {
-    font-size: 3em;
-}
-.sub-title {
-     font-size: 2em;
- }
