+++ src/app/core/models/app-config.ts
@@ -1,9 +1,5 @@
 import { Injectable } from '@angular/core';
 
-/**
- * Abstracts away application configuration.
- * TODO: This will be loaded from an external configuration file once we look at it
- */
 @Injectable()
 export class AppConfig {
