+++ src/app/core/services/configuration-service/index.ts
@@ -1,0 +1,1 @@
+export * from './configuration-service';
