+++ src/app/core/services/logging-service/index.ts
@@ -1,0 +1,2 @@
+export * from './logging.service';
+export * from './logging-error-handler';
