+++ src/app/features/app/app.e2e.ts
@@ -8,15 +8,9 @@ describe('AppComponent', function () {
 
 	it('should have a title', () => {
     let subject = browser.getTitle();
-    let result = 'StepChange Angular 2 Application Seed';
+    let result = 'StepChange Deadpool';
 
     expect(subject).toEqual(result);
   });
-
-	it('can load an image from the assets folder', () => {
-		let src = element(by.tagName('img')).getAttribute('src');
-		
-		expect(src).toContain('logo.jpg');
-	});
 	
 })
