+++ src/app/features/master-header/index.ts
@@ -1,0 +1,1 @@
+export * from './master-header.component';
