+++ src/index.html
@@ -2,7 +2,6 @@
 
 <head>
 	<title><%= webpackConfig.metadata.title %></title>
-	<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
 </head>
 
 <body>
