+++ src/theme/styles.scss
@@ -1,2 +1,1 @@
-@import "vendor/normalize.css";
-@import "vendor/skeleton.css";
+@import "/vendor/stepchange/stepchange.scss";
